move make_request function to utils module
